requirements.txt
python-dotenv==1.0.1
aiogram==3.15.0
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.27.2
